+++ nautilus_trader/adapters/binance/futures/execution.py
@@ -46,7 +46,7 @@ from nautilus_trader.model.enums import OrderType
 from nautilus_trader.model.enums import order_type_to_str
 from nautilus_trader.model.enums import time_in_force_to_str
 from nautilus_trader.model.identifiers import InstrumentId
-from nautilus_trader.model.orders.base import Order
+from nautilus_trader.model.orders import Order
 from nautilus_trader.msgbus.bus import MessageBus