+++ nautilus_trader/examples/strategies/ema_cross.py
@@ -29,9 +29,9 @@ from nautilus_trader.model.data.ticker import Ticker
 from nautilus_trader.model.enums import OrderSide
 from nautilus_trader.model.identifiers import InstrumentId
 from nautilus_trader.model.instruments import Instrument
-from nautilus_trader.model.orderbook.book import OrderBook
-from nautilus_trader.model.orderbook.data import OrderBookData
-from nautilus_trader.model.orders.market import MarketOrder
+from nautilus_trader.model.orderbook import OrderBook
+from nautilus_trader.model.orderbook import OrderBookData
+from nautilus_trader.model.orders import MarketOrder
 from nautilus_trader.trading.strategy import Strategy